Rename "Simplex" to "PointedAffineSpace".
This is a more accurate description of this class, especially when considering the involvement of a
distance function. The distance to a simplex is not the distance to (any of) its parent hyperplane(s).
However, we do not need the distance to the simplex itself - we only actually use the distance to the
parent hyperplane.